[WIP] Restructure SqlDestructiveChangesChecker
The purpose of this commit is to separate pure checks from IO. This
serves multiple purposes:

- It makes it easier to reuse the results of previous checks, e.g. to
avoid checking multiple times that there are non-null values in some
column.
- It makes it easier to provide alternative warnings when the database
we want to migrate is not accessible, or the checks time out. (e.g.
"This migration can be destructive if there are non-null values in such
and such column"). In other terms, we can do destructive change checks
without IO.
- It makes the code clearer, because checks have to be explicit about
what information they need about the current state of the database.